refactor: consolidate settings section components into tab components
Remove PersonasSection, EnvVarsSection, and StreamActionsSection components
and inline their logic directly into their parent tab components. This
simplifies the component hierarchy and removes a level of unnecessary
nesting. Update SettingsPage to use lazy-loaded tab components instead.
