--- frontend/src/components/settings/sections/EnvVarsSection.tsx
@@ -1,50 +1,0 @@
-import { Suspense } from 'react';
-import { useSettingsContext } from '@/hooks/useSettingsContext';
-import { useCrudForm } from '@/hooks/useCrudForm';
-import { createDefaultEnvVarForm, validateEnvVarForm } from '@/utils/settings';
-import { lazyNamed } from '@/utils/lazyNamed';
-
-const EnvVarsSettingsTab = lazyNamed(
-  () => import('@/components/settings/tabs/EnvVarsSettingsTab'),
-  'EnvVarsSettingsTab',
-);
-const EnvVarDialog = lazyNamed(
-  () => import('@/components/settings/dialogs/EnvVarDialog'),
-  'EnvVarDialog',
-);
-
-export function EnvVarsSection() {
-  const { localSettings, persistSettings } = useSettingsContext();
-
-  const envVarCrud = useCrudForm(localSettings, persistSettings, {
-    createDefault: createDefaultEnvVarForm,
-    validateForm: (form, editingIndex) =>
-      validateEnvVarForm(form, editingIndex, localSettings.custom_env_vars ?? []),
-    getArrayKey: 'custom_env_vars',
-    itemName: 'environment variable',
-  });
-
-  return (
-    <>
-      <EnvVarsSettingsTab
-        envVars={localSettings.custom_env_vars}
-        onAddEnvVar={envVarCrud.handleAdd}
-        onEditEnvVar={envVarCrud.handleEdit}
-        onDeleteEnvVar={envVarCrud.handleDelete}
-      />
-      {envVarCrud.isDialogOpen && (
-        <Suspense fallback={null}>
-          <EnvVarDialog
-            isOpen={envVarCrud.isDialogOpen}
-            isEditing={envVarCrud.editingIndex !== null}
-            envVar={envVarCrud.form}
-            error={envVarCrud.formError}
-            onClose={envVarCrud.handleDialogClose}
-            onSubmit={envVarCrud.handleSave}
-            onEnvVarChange={envVarCrud.handleFormChange}
-          />
-        </Suspense>
-      )}
-    </>
-  );
-}
