+++ frontend/src/components/settings/tabs/PersonasSettingsTab.tsx
@@ -1,48 +1,71 @@
-import { ListManagementTab } from '@/components/ui/ListManagementTab';
-import type { Persona } from '@/types/user.types';
+import { Suspense } from 'react';
 import { UserCircle } from 'lucide-react';
+import { ListManagementTab } from '@/components/ui/ListManagementTab';
+import { useSettingsContext } from '@/hooks/useSettingsContext';
+import { useCrudForm } from '@/hooks/useCrudForm';
+import { createDefaultPersonaForm, validatePersonaForm } from '@/utils/settings';
+import { lazyNamed } from '@/utils/lazyNamed';
+import type { Persona } from '@/types/user.types';
 
-interface PersonasSettingsTabProps {
-  personas: Persona[] | null;
-  onAddPersona: () => void;
-  onEditPersona: (index: number) => void;
-  onDeletePersona: (index: number) => void | Promise<void>;
-}
+const PersonaEditDialog = lazyNamed(
+  () => import('@/components/settings/dialogs/PersonaEditDialog'),
+  'PersonaEditDialog',
+);
 
-export const PersonasSettingsTab: React.FC<PersonasSettingsTabProps> = ({
-  personas,
-  onAddPersona,
-  onEditPersona,
-  onDeletePersona,
-}) => {
+export const PersonasSettingsTab: React.FC = () => {
+  const { localSettings, persistSettings } = useSettingsContext();
+
+  const personaCrud = useCrudForm(localSettings, persistSettings, {
+    createDefault: createDefaultPersonaForm,
+    validateForm: (form, editingIndex) =>
+      validatePersonaForm(form, editingIndex, localSettings.personas ?? []),
+    getArrayKey: 'personas',
+    itemName: 'persona',
+  });
+
   return (
-    <ListManagementTab<Persona>
-      title="Personas"
-      description="Create personas with custom system prompts. Select from the persona dropdown in the input bar. Claude, Codex, and OpenCode support replacing the system prompt; the persona selector is hidden for Cursor and Copilot."
-      items={personas}
-      emptyIcon={UserCircle}
-      emptyText="No personas configured yet"
-      emptyButtonText="Create Your First Persona"
-      addButtonText="Add Persona"
-      deleteConfirmTitle="Delete Persona"
-      deleteConfirmMessage={(persona) =>
-        `Are you sure you want to delete "${persona.name}"? This action cannot be undone.`
-      }
-      getItemKey={(persona) => persona.name}
-      onAdd={onAddPersona}
-      onEdit={onEditPersona}
-      onDelete={onDeletePersona}
-      renderItem={(persona) => (
-        <>
-          <h3 className="mb-2 truncate text-xs font-medium text-text-primary dark:text-text-dark-primary">
-            {persona.name}
-          </h3>
-          <p className="line-clamp-3 font-mono text-2xs leading-relaxed text-text-quaternary dark:text-text-dark-quaternary">
-            {persona.content}
-          </p>
-        </>
+    <>
+      <ListManagementTab<Persona>
+        title="Personas"
+        description="Create personas with custom system prompts. Select from the persona dropdown in the input bar. Claude, Codex, and OpenCode support replacing the system prompt; the persona selector is hidden for Cursor and Copilot."
+        items={localSettings.personas}
+        emptyIcon={UserCircle}
+        emptyText="No personas configured yet"
+        emptyButtonText="Create Your First Persona"
+        addButtonText="Add Persona"
+        deleteConfirmTitle="Delete Persona"
+        deleteConfirmMessage={(persona) =>
+          `Are you sure you want to delete "${persona.name}"? This action cannot be undone.`
+        }
+        getItemKey={(persona) => persona.name}
+        onAdd={personaCrud.handleAdd}
+        onEdit={personaCrud.handleEdit}
+        onDelete={personaCrud.handleDelete}
+        renderItem={(persona) => (
+          <>
+            <h3 className="mb-2 truncate text-xs font-medium text-text-primary dark:text-text-dark-primary">
+              {persona.name}
+            </h3>
+            <p className="line-clamp-3 font-mono text-2xs leading-relaxed text-text-quaternary dark:text-text-dark-quaternary">
+              {persona.content}
+            </p>
+          </>
+        )}
+        logContext="PersonasSettingsTab"
+      />
+      {personaCrud.isDialogOpen && (
+        <Suspense fallback={null}>
+          <PersonaEditDialog
+            isOpen={personaCrud.isDialogOpen}
+            isEditing={personaCrud.editingIndex !== null}
+            persona={personaCrud.form}
+            error={personaCrud.formError}
+            onClose={personaCrud.handleDialogClose}
+            onSubmit={personaCrud.handleSave}
+            onPersonaChange={personaCrud.handleFormChange}
+          />
+        </Suspense>
       )}
-      logContext="PersonasSettingsTab"
-    />
+    </>
   );
 };
